Extract main frame's file related logic.
